backend/requirements.txt
alembic==1.7.5

anyio==3.5.0

asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==37.0.2
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.4
Werkzeug==2.0.3
zipp==3.7.0
